+++ code/src/HISTORY.md
@@ -1,11 +1,12 @@
 
 
-**1.0.X**
+**1.0.2**
 - Support of internet password type with a variable length.
 - Support of password online checks.
 - Support of an additional summary window.
 - YAGNI code removed.
 - Bug fixing.
+- Copyright year and version number changed.
 
 **1.0.1**
 - Support of custom password phrases for WEP key generation.